order

+++ models/order_item.go
@@ -6,5 +6,5 @@ type OrderItem struct {
 	BaseModel
 	OrderId   uuid.UUID `json:"order_id"`
 	ProductId uuid.UUID `json:"product_id"`
-	Quantity  float64   `json:"quantity"`
+	Quantity  int       `json:"quantity"`
 }
